Move os testes de objeto Transient para o pacote correto
