+++ src/br/com/caelum/financas/modelo/estado/transiente/TestaTransientComSomenteUmPersistent.java
@@ -1,4 +1,4 @@
-package br.com.caelum.financas.modelo.estado;
+package br.com.caelum.financas.modelo.estado.transiente;
 
 import javax.persistence.EntityManager;
 
@@ -6,12 +6,21 @@ import br.com.caelum.financas.infra.JPAUtil;
 import br.com.caelum.financas.modelo.Conta;
 
 /*
- 
- Neste teste persistimos um objeto com o mesmo manager mas com transacoes diferentes
- 
+  Neste teste persistimos uma conta de forma simples, com somente:
+  - Criamos uma conta
+  - Persistimos a conta pelo manager
+  - O Hibernate gera um insert:
+  
+Hibernate: 
+    insert 
+    into
+        Conta
+        (agencia, numero, titular) 
+    values
+        (?, ?, ?)  
  */
 
-public class TestaTransientComDuasTransactions {
+public class TestaTransientComSomenteUmPersistent {
 
 	public static void main(String[] args) {
 		EntityManager manager = new JPAUtil().getEntityManager();
@@ -19,10 +28,6 @@ public class TestaTransientComDuasTransactions {
 		
 		manager.getTransaction().begin();
 		manager.persist(conta);
-		manager.getTransaction().commit();
-		
-		manager.getTransaction().begin();
-		conta.setTitular("Gama");
 		manager.getTransaction().commit();
 		
 		manager.close();
